Reorganización del directorio componentes
