+++ vitest.config.ts
@@ -10,7 +10,7 @@ export default defineConfig({
     ],
     coverage: {
       provider: 'istanbul',
-      reporter: 'lcov',
+      // reporter: 'lcov',
     },
   },
 });
